fix hierarchical foreground style modifier
